enhance granular synthesis by fixing grain step computation and pre-computing grains per samples decrease minimum grain duration add user defined map of sample frequency through filename pattern ## fix issue when no samples fix issue with silent additive synthesis when grains are setup

+++ samples.h
@@ -9,7 +9,7 @@
         uint32_t frames;
         unsigned int chn;
         unsigned int chn_m1;
-        float pitch; // hz
+        double pitch; // hz
         int samplerate;
     };
